Fixed Parts CES

+++ SWAS/Resource.cpp
@@ -130,7 +130,7 @@ void Resource::RestoreResource()
 ////  GETTERS AND SETTERS  ////
 ///////////////////////////////
 
-void Resource::SetResourceCount(int resourceCount)
+void Resource::SetResourceCount(double resourceCount)
 {
 	_resourceCount = resourceCount;
 }
